Add logic

+++ src/main/java/com/nilskulawiak/jetlagtracker/station/Station.java
@@ -28,8 +28,8 @@ public class Station {
     private String name;
 
     @Column(nullable = false)
-    private Double xCoordinate;
+    private Integer xCoordinate;
 
     @Column(nullable = false)
-    private Double yCoordinate;
+    private Integer yCoordinate;
 }
